pyproject.toml
[project]
name = "doc-flesh"
version = "0.1.0"
description = "This tool allows hosting canonical config files in HOME directory and syncing those to ALL repositories based on doc-skeleton locally."
readme = "README.md"
authors = [
    { name = "Jani Sourander", email = "jani.sourander@kamk.fi" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "gitpython>=3.1.44",
    "jinja2>=3.1.6",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
]

[project.scripts]
doc-flesh = "doc_flesh.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--verbose"] # Add "--capture=no" to see print statements for debugging.
# Alternative: run only one test with no-stdout-capture:
# $ uv run pytest -s tests/test_some_specific_file.py::test_some_specific_function
